+++ src/odr/internal/pdf/pdf_document_parser.hpp
@@ -110,13 +110,6 @@ private:
   /// table together with the newest (first-seen) trailer dictionary.
   [[nodiscard]] std::pair<Xref, Dictionary> read_trailer_chain();
 
-  /// Last-resort cross-reference recovery for broken files (missing/garbage
-  /// `startxref`, wrong offsets, a damaged chain): forward-scan the whole file
-  /// for `n g obj` starts, rebuilding `m_xref` (last definition of an id wins)
-  /// and collecting `trailer` dictionaries into `m_trailer`. Then object-stream
-  /// members are indexed (`index_object_streams`) and, if no `trailer` supplied
-  /// a `/Root`, a `/Type /Catalog` object is searched (`recover_root`). Sets
-  /// `m_recovered`. Any object cached from the failed attempt is dropped first.
   void recover_xref();
   /// Index the members of every recovered `/Type /ObjStm` object as compressed
   /// cross-reference entries (additive; an existing direct entry wins).
@@ -127,10 +120,6 @@ private:
 
   [[nodiscard]] const ObjectStream &
   load_object_stream(const ObjectReference &reference);
-
-  /// Decrypt every string leaf of `object` in place with the owning object's
-  /// reference (ISO 32000-1 7.6.2). Used on freshly read indirect objects.
-  void decrypt_strings(Object &object, const ObjectReference &reference);
 
   std::unique_ptr<std::istream> m_stream;
   FileParser m_parser;
